More misc fixes -syntax -casting

+++ a3/a3-server-jhou-shsu/src/main/java/ca/bcit/infosys/a3/server/domain/Result.java
@@ -75,7 +75,7 @@ public class Result {
         return totalPossibleScore;
     }
 
-    public void setTotalPossibleScore(short totalPossibleScore) {
+    public void setTotalPossibleScore(int totalPossibleScore) {
         this.totalPossibleScore = totalPossibleScore;
     }
 }
